Implement average charge state equilibria.
Electron temperature dependent. Uses a polynomial fit for a selection of common impurities based on Mavrin 2018, including heavy impurities.

* Adds a new routine to calculate average impurity charge state for an IonMixture.

* Modifies the DynamicIonMixture dataclass to hold a separate Z_override attribute instead of calculating the average itself. This is used in the charge state calculation function.

* Updates core_profile_setters to calculate and return the Zimp/Zi profiles according to the new routines

* Better separation of ne and ion density + charge state setting

* Updates Qualikiz to use the new Zimp/Zi arrays

* Adds average Zimp to default output plots

* Updates boundary condition calculations to calculate boundary Zimp according to the boundary temperature.

* New sim integration test with tungsten added.

* Some sim tests which did not have a Z_override now have spatially varying Zimp, which leads to small O(1e-4) changes. The small impact is because the default impurity is Neon.

Small <5% runtime increase. This was ascertained to not be due to the (jitted) charge state calculations themselves, but due to surrounding code in non-jitted functions in the sim loop which call the charge state calculations. This will be rectified in an upcoming PR.

~10% compilation time increase due to added dependency of nimp and zimp on electron temperature.

PiperOrigin-RevId: 717475094

+++ torax/constants.py
@@ -39,7 +39,7 @@ class IonProperties:
   symbol: str
   name: str
   A: float
-  Z: int
+  Z: float
 
 
 @chex.dataclass(frozen=True)
@@ -66,21 +66,21 @@ CONSTANTS: Final[Constants] = Constants(
 # Taken from
 # https://www.nist.gov/pml/periodic-table-elements and https://ciaaw.org.
 ION_PROPERTIES: Final[tuple[IonProperties, ...]] = (
-    IonProperties(symbol='H', name='Hydrogen', A=1.008, Z=1),
-    IonProperties(symbol='D', name='Deuterium', A=2.0141, Z=1),
-    IonProperties(symbol='T', name='Tritium', A=3.0160, Z=1),
-    IonProperties(symbol='He3', name='Helium-3', A=3.0160, Z=2),
-    IonProperties(symbol='He4', name='Helium-4', A=4.0026, Z=2),
-    IonProperties(symbol='Li', name='Lithium', A=5.3917, Z=3),
-    IonProperties(symbol='Be', name='Beryllium', A=9.0122, Z=4),
-    IonProperties(symbol='C', name='Carbon', A=12.011, Z=6),
-    IonProperties(symbol='N', name='Nitrogen', A=14.007, Z=7),
-    IonProperties(symbol='O', name='Oxygen', A=15.999, Z=8),
-    IonProperties(symbol='Ne', name='Neon', A=20.180, Z=10),
-    IonProperties(symbol='Ar', name='Argon', A=39.95, Z=18),
-    IonProperties(symbol='Kr', name='Krypton', A=83.798, Z=36),
-    IonProperties(symbol='Xe', name='Xenon', A=131.29, Z=54),
-    IonProperties(symbol='W', name='Tungsten', A=183.84, Z=74),
+    IonProperties(symbol='H', name='Hydrogen', A=1.008, Z=1.0),
+    IonProperties(symbol='D', name='Deuterium', A=2.0141, Z=1.0),
+    IonProperties(symbol='T', name='Tritium', A=3.0160, Z=1.0),
+    IonProperties(symbol='He3', name='Helium-3', A=3.0160, Z=2.0),
+    IonProperties(symbol='He4', name='Helium-4', A=4.0026, Z=2.0),
+    IonProperties(symbol='Li', name='Lithium', A=5.3917, Z=3.0),
+    IonProperties(symbol='Be', name='Beryllium', A=9.0122, Z=4.0),
+    IonProperties(symbol='C', name='Carbon', A=12.011, Z=6.0),
+    IonProperties(symbol='N', name='Nitrogen', A=14.007, Z=7.0),
+    IonProperties(symbol='O', name='Oxygen', A=15.999, Z=8.0),
+    IonProperties(symbol='Ne', name='Neon', A=20.180, Z=10.0),
+    IonProperties(symbol='Ar', name='Argon', A=39.95, Z=18.0),
+    IonProperties(symbol='Kr', name='Krypton', A=83.798, Z=36.0),
+    IonProperties(symbol='Xe', name='Xenon', A=131.29, Z=54.0),
+    IonProperties(symbol='W', name='Tungsten', A=183.84, Z=74.0),
 )
 
 ION_PROPERTIES_DICT: Final[Mapping[str, IonProperties]] = (
